fix: 修复删除的时候开源版报错的问题

+++ frontend/src/business/components/track/case/components/TestCaseList.vue
@@ -901,7 +901,14 @@ export default {
         // 删除提供列表删除和全部版本删除
         this.$refs.apiDeleteConfirm.open(testCase, this.$t('test_track.case.delete_confirm'));
       } else {
-        this._handleDeleteVersion(testCase, false);
+        this.$alert(this.$t('test_track.case.delete_confirm') + '\'' + testCase.name + '\'' + "？", '', {
+          confirmButtonText: this.$t('commons.confirm'),
+          callback: (action) => {
+            if (action === 'confirm') {
+              this._handleDeleteVersion(testCase, false);
+            }
+          }
+        });
       }
     },
     batchReduction() {
